+++ enum/src/main/java/org/grnet/status/enums/IncidentStatus.java
@@ -1,0 +1,10 @@
+package org.grnet.status.enums;
+
+public enum IncidentStatus {
+
+    REPORTED,
+    INVESTIGATING,
+    IDENTIFIED,
+    WATCHING,
+    CLOSED
+}